+++ src/main/java/nextstep/exception/DuplicateEntityException.java
@@ -1,4 +1,4 @@
-package nextstep.support;
+package nextstep.exception;
 
 public class DuplicateEntityException extends RuntimeException {
     public DuplicateEntityException() {
