+++ src/main/java/nextstep/exception/DuplicateReservationException.java
@@ -1,4 +1,4 @@
-package nextstep.support;
+package nextstep.exception;
 
 public class DuplicateReservationException extends DuplicateEntityException {
     public DuplicateReservationException() {
