+++ src/main/java/nextstep/exception/InvalidAuthorizationTokenException.java
@@ -1,4 +1,4 @@
-package nextstep.support;
+package nextstep.exception;
 
 public class InvalidAuthorizationTokenException extends RuntimeException {
     public InvalidAuthorizationTokenException() {
