+++ src/main/java/nextstep/exception/NotCorrectPasswordException.java
@@ -1,4 +1,4 @@
-package nextstep.support;
+package nextstep.exception;
 
 public class NotCorrectPasswordException extends RuntimeException {
     public NotCorrectPasswordException() {
