+++ src/main/java/nextstep/exception/NotExistMemberException.java
@@ -1,4 +1,4 @@
-package nextstep.support;
+package nextstep.exception;
 
 public class NotExistMemberException extends NotExistEntityException {
     public NotExistMemberException() {
