+++ src/main/java/nextstep/exception/NotExistReservationException.java
@@ -1,4 +1,4 @@
-package nextstep.support;
+package nextstep.exception;
 
 public class NotExistReservationException extends NotExistEntityException {
     public NotExistReservationException() {
