+++ src/main/java/nextstep/exception/NotQualifiedMemberException.java
@@ -1,4 +1,4 @@
-package nextstep.support;
+package nextstep.exception;
 
 public class NotQualifiedMemberException extends RuntimeException {
     public NotQualifiedMemberException() {
