+++ src/main/java/nextstep/support/LoginMemberArgumentResolver.java
@@ -2,7 +2,8 @@ package nextstep.support;
 
 import nextstep.auth.AuthorizationTokenExtractor;
 import nextstep.auth.JwtTokenProvider;
-import nextstep.member.Member;
+import nextstep.exception.InvalidAuthorizationTokenException;
+import nextstep.exception.NotExistMemberException;
 import nextstep.member.MemberDao;
 import org.springframework.core.MethodParameter;
 import org.springframework.stereotype.Component;
@@ -10,8 +11,6 @@ import org.springframework.web.bind.support.WebDataBinderFactory;
 import org.springframework.web.context.request.NativeWebRequest;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.method.support.ModelAndViewContainer;
-
-import java.util.Optional;
 
 @Component
 public class LoginMemberArgumentResolver implements HandlerMethodArgumentResolver {
